auth/requirements.txt
flask==2.3.2
gunicorn==20.1.0
pyjwt==2.7.0
requests
